validate raster grid coordinates when running x ray centering
