+++ mx3_beamline_library/devices/classes/motors.py
@@ -501,7 +501,7 @@ class MD3Motor(Signal):
         -------
         None
         """
-        limits = self.get_limits
+        limits = self.limits
         if not limits[0] <= value <= limits[1]:
             raise ValueError(
                 f"Value is out of limits. Given value was {value}, "
@@ -531,7 +531,7 @@ class MD3Motor(Signal):
             self.server.setMotorPosition(self.name, value)
 
     @cached_property
-    def get_limits(self) -> tuple[float, float]:
+    def limits(self) -> tuple[float, float]:
         return tuple(self.server.getMotorDynamicLimits(self.name))
 
     def get(self) -> float:
@@ -931,7 +931,7 @@ class MD3PLateTranslation(Signal):
         -------
         None
         """
-        limits = self.get_limits
+        limits = self.limits
         if not limits[0] <= value <= limits[1]:
             raise ValueError(
                 f"Value is out of limits. Given value was {value}, "
@@ -960,7 +960,7 @@ class MD3PLateTranslation(Signal):
         return self.get()
 
     @cached_property
-    def get_limits(self) -> tuple[float, float]:
+    def limits(self) -> tuple[float, float]:
         return tuple(self.server.getMotorDynamicLimits(self.name))
